Disable NDEBUG in builds
In principle systemd supports building without assertions for production,
but we want the assertions to be enabled to catch as many errors early as possible.

Also, let's remove the obsolete work-around for meson not showing logs. This
is already reverted upstream, but apparently not in the version of macros that
Fedora has.