Conditionalize dracut Conflicts more
Make sure on centos stream 10 we also conflict with dracut 060-2
and that on centos stream 9 so that the spec can still be used to
build systemd rpms for centos stream 9 upstream in systemd CI that
can be installed on centos stream 9.

(dracut is pulled in as a required dependency of kernel-core so we
can't just not install it on centos stream 9 unfortunately).